Move "safe" and "finalized" handling to merge plugin
Chains like ETC and Polygon don't support safe and finalized, so
there's no reason to be running those queries in the core producer
plugin when the merge plugin should be present on all the systems
with safe / finalized to take into consideration.
